增加点检项属性和接口参数检查

+++ src/main/java/me/cuiyijie/joyea/controller/ProductController.java
@@ -1,6 +1,7 @@
 package me.cuiyijie.joyea.controller;
 
 import com.github.pagehelper.Page;
+import com.google.common.collect.Lists;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import me.cuiyijie.joyea.model.Product;
@@ -10,6 +11,9 @@ import me.cuiyijie.joyea.pojo.TransBaseResponse;
 import me.cuiyijie.joyea.pojo.TransProductRequest;
 import me.cuiyijie.joyea.pojo.TransProjectRequest;
 import me.cuiyijie.joyea.service.ProductService;
+import me.cuiyijie.joyea.util.CheckParamsUtil;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -18,14 +22,18 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.List;
+
 /**
- * @Author: yjcui3
+ * @Author: cuiyijie
  * @Date: 2021/10/18 11:25
  */
 @RestController
 @RequestMapping("product")
 @Api(tags = "项目产品模块")
 public class ProductController {
+
+    private Logger logger = LoggerFactory.getLogger(ProductController.class);
 
     @Autowired
     private ProductService productService;
@@ -48,7 +56,14 @@ public class ProductController {
     @RequestMapping(value = "insert", method = RequestMethod.POST)
     public TransBaseResponse insert(@RequestBody TransProductRequest request) {
         TransBaseResponse response = new TransBaseResponse();
-
+        List<String> paramsCheck = Lists.newArrayList("projectId:项目id（projectId）","productName:产品名称（productName）", "productNumber:产品编号（productNumber）");
+        String errorMsg = CheckParamsUtil.checkAll(request, paramsCheck, null, null);
+        if (errorMsg != null) {
+            logger.error("参数检查错误：" + errorMsg);
+            response.setCode("0");
+            response.setMsg(errorMsg);
+            return response;
+        }
         Product selection = new Product();
         BeanUtils.copyProperties(request, selection);
         Integer result = productService.insert(selection);
@@ -64,6 +79,14 @@ public class ProductController {
     @RequestMapping(value = "delete", method = RequestMethod.POST)
     public TransBaseResponse delete(@RequestBody TransProductRequest request) {
         TransBaseResponse response = new TransBaseResponse();
+        List<String> paramsCheck = Lists.newArrayList("id:产品id（id）");
+        String errorMsg = CheckParamsUtil.checkAll(request, paramsCheck, null, null);
+        if (errorMsg != null) {
+            logger.error("参数检查错误：" + errorMsg);
+            response.setCode("0");
+            response.setMsg(errorMsg);
+            return response;
+        }
         Integer result = productService.delete(request.getId());
         if (result == 1) {
             response.setCode("0");
@@ -77,6 +100,14 @@ public class ProductController {
     @RequestMapping(value = "update", method = RequestMethod.POST)
     public TransBaseResponse update(@RequestBody TransProductRequest request) {
         TransBaseResponse response = new TransBaseResponse();
+        List<String> paramsCheck = Lists.newArrayList("id:产品id（id）","productName:产品名称（productName）", "productNumber:产品编号（productNumber）");
+        String errorMsg = CheckParamsUtil.checkAll(request, paramsCheck, null, null);
+        if (errorMsg != null) {
+            logger.error("参数检查错误：" + errorMsg);
+            response.setCode("0");
+            response.setMsg(errorMsg);
+            return response;
+        }
         Product selection = new Product();
         BeanUtils.copyProperties(request,selection);
         Integer result = productService.update(selection);
